api/requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coverage==4.5.2
cryptography==2.4.2
Django==2.1
django-rest-knox==3.1.5
djangorestframework==3.8.2
idna==2.8
model-mommy==1.6.0
pycparser==2.19
pyOpenSSL==18.0.0
pytz==2018.5
requests==2.21.0
six==1.12.0
urllib3==1.24.1
